requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.5.0
matplotlib>=3.7.0
tqdm>=4.65.0
pyyaml>=6.0
scipy>=1.10.0
lpips>=0.1.4
scikit-image>=0.21.0
requests>=2.31.0
pandas>=2.0.0
seaborn>=0.12.0
